+++ src/main/kotlin/ru/vdnh/repository/RouteRepository.kt
@@ -1,9 +1,8 @@
 package ru.vdnh.repository
 
 import ru.vdnh.model.entity.RouteEntity
-import java.math.BigInteger
 
 interface RouteRepository {
 
-    fun getRouteById(id: BigInteger): RouteEntity
+    fun getRouteById(id: Long): RouteEntity
 }
